Convert to Static Lib and Fix warnings

+++ MawarEngine/src/Mawar/ImGui/ImGuiLayer.cpp
@@ -97,7 +97,7 @@ namespace Mawar
 		Application& app = Application::Get();
 		float time = (float)glfwGetTime();
 
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 		io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
 		m_Time = time;
